+++ llama-index-integrations/indices/llama-index-indices-managed-vectara/llama_index/indices/managed/vectara/base.py
@@ -102,13 +102,19 @@ class VectaraIndex(BaseManagedIndex):
 
         # setup requests session with max 3 retries and 90s timeout
         # for calling Vectara API
-        self._session = requests.Session()  # to reuse connections
+        self._session = requests.Session()
         if not vectara_verify_ssl:
             self._session.verify = False  # to ignore SSL verification
         adapter = requests.adapters.HTTPAdapter(max_retries=3)
         self._session.mount("https://", adapter)
         self.vectara_api_timeout = 90
         self.doc_ids: List[str] = []
+
+    def __del__(self) -> None:
+        """Attempt to close the session when the object is garbage collected."""
+        if hasattr(self, "_session") and self._session:
+            self._session.close()
+            self._session = None
 
     @lru_cache(maxsize=None)
     def _get_corpus_key(self, corpus_key: str) -> str: