+++ llama-index-integrations/llms/llama-index-llms-openai/llama_index/llms/openai/utils.py
@@ -46,6 +46,10 @@ O1_MODELS: Dict[str, int] = {
     "o1-mini-2024-09-12": 128000,
     "o3-mini": 200000,
     "o3-mini-2025-01-31": 200000,
+    "o3": 200000,
+    "o3-2025-04-16": 200000,
+    "o4-mini": 200000,
+    "o4-mini-2025-04-16": 200000,
 }
 
 O1_MODELS_WITHOUT_FUNCTION_CALLING = {
@@ -92,6 +96,13 @@ GPT4_MODELS: Dict[str, int] = {
     # 0314 models
     "gpt-4-0314": 8192,
     "gpt-4-32k-0314": 32768,
+    # GPT 4.1 Models
+    "gpt-4.1": 1047576,
+    "gpt-4.1-mini": 1047576,
+    "gpt-4.1-nano": 1047576,
+    "gpt-4.1-2025-04-14": 1047576,
+    "gpt-4.1-mini-2025-04-14": 1047576,
+    "gpt-4.1-nano-2025-04-14": 1047576,
 }
 
 AZURE_TURBO_MODELS: Dict[str, int] = {
@@ -257,6 +268,10 @@ def is_chat_model(model: str) -> bool:
 
 
 def is_function_calling_model(model: str) -> bool:
+    # default to True for models that are not in the ALL_AVAILABLE_MODELS dict
+    if model not in ALL_AVAILABLE_MODELS:
+        return True
+
     # checking whether the model is fine-tuned or not.
     # fine-tuned model names these days look like:
     # ft:gpt-3.5-turbo:acemeco:suffix:abc123
@@ -473,7 +488,7 @@ def to_openai_responses_message_dict(
             "role": message.role.value,
             "content": (
                 content_txt
-                if message.role.value in ("assistant", "system", "developer")
+                if message.role.value in ("system", "developer")
                 or all(isinstance(block, TextBlock) for block in message.blocks)
                 else content
             ),
@@ -510,7 +525,7 @@ def to_openai_message_dicts(
             message_dicts = to_openai_responses_message_dict(
                 message,
                 drop_none=drop_none,
-                model=model,
+                model="o3-mini",  # hardcode to ensure developer messages are used
             )
             if isinstance(message_dicts, list):
                 final_message_dicts.extend(message_dicts)