+++ llama-index-integrations/vector_stores/llama-index-vector-stores-milvus/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [v0.8.0]
+
+### Changed
+
+- Changed the default sparse embedding function from `BGEM3SparseEmbeddingFunction` to `BM25BuiltInFunction`.
+
 ## [v0.2.1]
 
 ### Changed